Refactor some typings
